server/requirements.txt
livekit-agents[assemblyai,openai,silero,rime]
livekit-plugins-noise-cancellation
python-dotenv
